Deal with PG secrets with the rest of the PG stuf
